moved ELM-specific data management from train_base.py to elm_data.py
