+++ bes_ml/base/__init__.py
@@ -1,2 +1,2 @@
-from .train_base import _Trainer_Base
+from .train_base import _Base_Trainer
 from .analyze_base import _Analyzer_Base
