+++ bes_ml/base/multisource_data.py
@@ -12,46 +12,6 @@ import torch
 import torch.utils.data
 
 
-# TODO: make dataclass
-class ELM_Dataset(torch.utils.data.Dataset):
-
-    def __init__(
-        self,
-        signals: np.ndarray = None, 
-        labels: np.ndarray = None, 
-        sample_indices: np.ndarray = None, 
-        window_start: np.ndarray = None,  # TODO: refactor to remove `window_start` parameter
-        signal_window_size: int = None,
-        prediction_horizon: int = None,  # =0 for time-to-ELM regression; >=0 for classification prediction
-    ) -> None:
-        self.signals = torch.unsqueeze(torch.from_numpy(signals), 0)
-        assert (
-            self.signals.ndim == 4 and 
-            self.signals.shape[0] == 1 and 
-            self.signals.shape[2] == 8 and 
-            self.signals.shape[3] == 8
-        ), "Signals have incorrect shape"
-        self.labels = torch.from_numpy(labels)
-        assert self.labels.ndim == 1, "Labels have incorrect shape"
-        assert self.labels.shape[0] == self.signals.shape[1], "Labels and signals have different time dimensions"
-        self.sample_indices = sample_indices
-        # self.window_start = window_start
-        self.signal_window_size = signal_window_size
-        self.prediction_horizon = prediction_horizon if prediction_horizon is not None else 0
-
-    def __len__(self):
-        return self.sample_indices.size
-
-    def __getitem__(self, idx: int):
-        time_idx = self.sample_indices[idx]
-        # BES signal window data
-        signal_window = self.signals[:, time_idx : time_idx + self.signal_window_size, :, :]
-        # label for signal window
-        label = self.labels[ time_idx + self.signal_window_size + self.prediction_horizon - 1 ]
-
-        return signal_window, label
-
-
 class MultiSourceDataset(torch.utils.data.Dataset):
 
     def __init__(self,
