+++ bes_ml/base/train_base.py
@@ -1,21 +1,15 @@
 # python library imports
 import contextlib
-import io
 import logging
-import pickle
-import sys
 import time
 from pathlib import Path
-from typing import Union, Iterable, Tuple, Any
+from typing import Union, Tuple, Any
 import dataclasses
 
 # 3rd-party imports
-import h5py
 import numpy as np
-import matplotlib.pyplot as plt
 import torch
 import torch.utils.data
-import torchinfo  # conda-forge
 import yaml
 from sklearn import metrics
 
@@ -25,63 +19,81 @@ except ImportError:
     optuna = None
 
 # repo import
-from bes_data.sample_data import sample_elm_data_file
 try:
-    from .models import Multi_Features_Model, _Multi_Features_Model_Dataclass
-    from .utilities import merge_pdfs
+    from .models import Multi_Features_Model
 except ImportError:
-    from bes_ml.base.models import Multi_Features_Model, _Multi_Features_Model_Dataclass
-    from bes_ml.base.utilities import merge_pdfs
+    from bes_ml.base.models import Multi_Features_Model
 
 
-@dataclasses.dataclass
-class _Trainer_Base(_Multi_Features_Model_Dataclass):
-    data_location: Union[Path, str] = sample_elm_data_file  # path to data; dir or file depending on task
-    output_dir: Union[Path,str] = Path('run_dir')  # path to output dir.
+@dataclasses.dataclass(eq=False)
+class _Base_Trainer_Dataclass:
+    # output parameters
+    output_dir: Path = Path('run_dir')  # path to output dir.
     results_file: str = 'results.yaml'  # output training results
     log_file: str = 'log.txt'  # output log file
     inputs_file: str = 'inputs.yaml'  # save inputs to yaml
-    # test_data_file: str = 'test_data.pkl'  # if None, do not save test data (can be large)
-    checkpoint_file: str = 'checkpoint.pytorch'  # pytorch save file; if None, do not save
-    data_partition_file: str = 'data_partition.yaml'  # data partition for training, valid., and testing
-    export_onnx: bool = False  # export ONNX format
+    checkpoint_file: str = 'checkpoint.pytorch'  # pytorch save file
+    save_onnx_model: bool = False  # export ONNX format
+    onnx_checkpoint_file: str = 'checkpoint.onnx'  # onnx save file
+    logger: logging.Logger = None
+    # training parameters
     device: str = 'auto'  # auto (default), cpu, cuda, or cuda:X
-    num_workers: int = 0  # number of subprocess workers for pytorch dataloader
     n_epochs: int = 2  # training epochs
-    batch_size: int = 64  # power of 2, like 16-128
-    minibatch_interval: int = 2000  # print minibatch info
-    signal_window_size: int = 128  # power of 2, like 32-512
-    fraction_validation: float = 0.1  # fraction of dataset for validation
-    fraction_test: float = 0.15  # fraction of dataset for testing
+    minibatch_print_interval: int = 2000  # print minibatch info
     optimizer_type: str = 'adam'  # adam (default) or sgd
     sgd_momentum: float = 0.0  # momentum for SGD optimizer
     sgd_dampening: float = 0.0  # dampening for SGD optimizer
     learning_rate: float = 1e-3  # optimizer learning rate
     lr_scheduler_patience: int = 4  # epochs to wait before triggering lr scheduler
     weight_decay: float = 5e-3  # optimizer L2 regularization factor
-    batches_per_print: int = 5000  # train/validation batches per print update
-    logger: logging.Logger = None
-    seed: int = None  # RNG seed for deterministic, reproducable shuffling (ELMs, sample indices, etc.)
-    trial: Any = None  # optuna trial
-    normalize_signals: bool = True  # if True, normalize BES signals such that max ~= 1
-    normalize_labels: bool = False  # if True, normalize regression labels to min/max = +/- 1
+    # optuna integration
+    optuna_trial: Any = None  # optuna trial
+    # global parameters
+    is_regression: bool = dataclasses.field(default=None, init=False)
+    is_classification: bool = dataclasses.field(default=None, init=False)
+    model: Multi_Features_Model = dataclasses.field(default=None, init=False)
+    train_data_loader: torch.utils.data.DataLoader = dataclasses.field(default=None, init=False)
+    validation_data_loader: torch.utils.data.DataLoader = dataclasses.field(default=None, init=False)
+    results: dict = dataclasses.field(default=None, init=False)
+    label_type: np.dtype = dataclasses.field(default=None, init=False)
+
+
+@dataclasses.dataclass(eq=False)
+class _Base_Trainer(_Base_Trainer_Dataclass):
+    """Base class for model trainer"""
 
     def __post_init__(self):
-        self.data_location = Path(self.data_location).resolve()
-        assert self.data_location.exists(), f"{self.data_location} does not exist"
-
         self.output_dir = Path(self.output_dir).resolve()
         self.output_dir.mkdir(exist_ok=True, parents=True)
-
-        # subclass must set either is_regression or is_classification to True
-        self.is_regression = False
-        self.is_classification = False
-
-        self.rng_generator = np.random.default_rng(seed=self.seed)
 
         self._create_logger()
         self._print_inputs()
         self._save_inputs_to_yaml()
+
+        self.results = {
+            'train_loss': [],
+            'valid_loss': [],
+            'loss_function_name': '',
+            'scores': [],
+            'score_function_name': '',
+            'training_time': [],
+            'lr': [],
+        }
+
+        # subclass must set is_regression XOR is_classification
+        assert self.is_regression ^ self.is_classification  # XOR
+
+        self._make_model()
+        self._set_device()
+        self.model = self.model.to(self.device)
+        self._set_regression_classification()
+        self._make_optimizer_scheduler()
+        self._prepare_data()
+
+        # validate data loaders
+        for data_loader in [self.train_data_loader, self.validation_data_loader]:
+            assert isinstance(data_loader, torch.utils.data.DataLoader) or data_loader is None
+
 
     def _create_logger(self):
         """
@@ -120,7 +132,7 @@ class _Trainer_Base(_Multi_Features_Model_Dataclass):
     def _save_inputs_to_yaml(self):
         filename = Path(self.output_dir / self.inputs_file)
         self_fields_dict = dataclasses.asdict(self)
-        for skip_key in ['logger', 'trial']:
+        for skip_key in ['logger', 'optuna_trial']:
             self_fields_dict.pop(skip_key)
         for key in self_fields_dict:
             if isinstance(self_fields_dict[key], Path):
@@ -133,369 +145,52 @@ class _Trainer_Base(_Multi_Features_Model_Dataclass):
                 sort_keys=False,
             )
 
-    def make_model_and_set_device(self):
-
-        # make model
+    def _make_model(self) -> None:
         model_kwargs = {
             field.name: getattr(self, field.name) 
             for field in dataclasses.fields(Multi_Features_Model)
         }
         self.model = Multi_Features_Model(**model_kwargs)
+        self.model.print_model_summary()
 
-        # setup device
+    def _set_device(self) -> None:
         if self.device == 'auto':
             self.device = 'cuda' if torch.cuda.is_available() else 'cpu'
         self.device = torch.device(self.device)
         self.logger.info(f"Device: {self.device}")
 
-        # send model to device
-        self.model = self.model.to(self.device)
+    def _set_regression_classification(self) -> None:
 
-        # print model summary
-        self.input_shape = (self.batch_size, 1, self.signal_window_size, 8, 8)
-        self._print_model_summary()
-
-    def _print_model_summary(self):
-        self.logger.info("MODEL SUMMARY")
-
-        # catpure torchinfo.summary() output
-        tmp_io = io.StringIO()
-        sys.stdout = tmp_io
-        print()
-        torchinfo.summary(self.model, input_size=self.input_shape, device=self.device)
-        sys.stdout = sys.__stdout__
-        self.logger.info(tmp_io.getvalue())
-        # print model summary
-        n_params = sum(p.numel() for p in self.model.parameters() if p.requires_grad)
-        self.logger.info(f"Model contains {n_params} trainable parameters")
-        test_input = torch.rand(*self.input_shape).to(self.device)
-        self.logger.info(f'Batched input size: {test_input.shape}')
-        test_output = self.model(test_input)
-        self.logger.info(f"Batched output size: {test_output.shape}")
-
-    def finish_subclass_initialization(self) -> None:
-        # subclass must set is_regression XOR is_classification
-        assert self.is_regression ^ self.is_classification  # XOR
         if self.is_regression:
-            # regression model (e.g. time to ELM onset)
-            self.loss_function = torch.nn.MSELoss(reduction="none")
-            self.score_function = metrics.r2_score
+            # regression model (e.g. time to ELM onset, velocimetry surrogate model)
+            self.label_type = np.float32
             self.score_function_name = 'R2'
+            self.loss_function_name = 'MSELoss'
         elif self.is_classification:
+            self.label_type = np.int8
+            self.score_function_name = 'F1'
             if self.model.mlp_output_size == 1:
-                # binary classification (e.g. active ELM?)
-                self.loss_function = torch.nn.BCEWithLogitsLoss(reduction="none")
+                # binary classification (e.g. active/inactive ELM)
+                self.loss_function_name = 'BCEWithLogitsLoss'
+                assert hasattr(self, 'threshold')  # binary classification must specify threshold
             else:
                 # multi-class classification (e.g. confinement mode)
-                self.loss_function = torch.nn.CrossEntropyLoss(reduction="none")
-            self.score_function = metrics.f1_score
-            self.score_function_name = 'F1'
-        assert (
-            self.loss_function and 
-            self.score_function and 
-            self.score_function_name
-        )
+                self.loss_function_name = 'CrossEntropyLoss'
 
-        self.optimizer = self.lr_scheduler = None
-        self._make_optimizer_scheduler_loss()
+        self.loss_function = getattr(torch.nn, self.loss_function_name)(reduction="none")
 
-        # get data
-        self.train_data = self.validation_data = self.test_data = None
-        self._get_data()
-        # ensure all data is np.ndarray at this point
-        for data_list in [self.train_data, self.validation_data, self.test_data]:
-            if data_list is None: continue
-            for item in data_list:
-                assert isinstance(item, np.ndarray)
-        
-        # make datasets
-        self.train_dataset = self.validation_dataset = None
-        self._make_datasets()
-        # ensure datasets are torch.utils.data.Dataset
-        for dataset in [self.train_dataset, self.validation_dataset]:
-            if dataset is None: continue
-            assert isinstance(dataset, torch.utils.data.Dataset)
+        self.results['score_function_name'] = self.score_function_name
+        self.results['loss_function_name'] = self.loss_function_name
 
-        # make data loaders
-        self.train_data_loader = self.validation_data_loader = None
-        self._make_data_loaders()
-        # ensure data loaders are torch.utils.data.DataLoader
-        for data_loader in [self.train_data_loader, self.validation_data_loader]:
-            if data_loader is None: continue
-            assert isinstance(data_loader, torch.utils.data.DataLoader)
-
-        # if self.test_data_file and self.fraction_test>0.0:
-        #     self._save_test_data()
-
-    # def _save_test_data(self) -> None:
-    #     """
-    #     Save test data to pickle file.
-    #     """
-    #     test_data_file = self.output_dir / self.test_data_file
-    #     self.logger.info(f"Test data file: {test_data_file}")
-    #     with test_data_file.open('wb') as file:
-    #         pickle.dump(
-    #             {
-    #                 "signals": self.test_data[0],
-    #                 "labels": self.test_data[1],
-    #                 "sample_indices": self.test_data[2],
-    #                 "window_start": self.test_data[3],
-    #                 "elm_indices": self.test_data[4],
-    #             },
-    #             file,
-    #         )
-    #     assert test_data_file.exists(), f"{test_data_file} does not exist"
-    #     self.logger.info(f"  File size: {test_data_file.stat().st_size/1e6:.1f} MB")
-
-    def _get_data(self) -> None:
-        self.data_location = self.data_location.resolve()
-        assert self.data_location.exists(), f"{self.data_location} does not exist"
-        self.logger.info(f"Data file: {self.data_location}")
-
-        with h5py.File(self.data_location, "r") as data_file:
-            elm_indices = np.array(
-                [int(key) for key in data_file], 
-                dtype=int,
-            )
-            time_frames = sum([data_file[key]['time'].shape[0] for key in data_file])
-        self.logger.info(f"Events in data file: {elm_indices.size}")
-        self.logger.info(f"Total time frames: {time_frames}")
-
-        # shuffle ELM events
-        self.rng_generator.shuffle(elm_indices)
-
-        # TODO: remove `max_elms` from base class
-        if hasattr(self, 'max_elms') and self.max_elms:
-            elm_indices = elm_indices[:self.max_elms]
-            self.logger.info(f"Limiting data to {self.max_elms} ELM events")
-
-        if elm_indices.size >= 5:
-            self.logger.info(f"Initial ELM indices: {elm_indices[0:5]}")
-
-        n_validation_elms = int(self.fraction_validation * elm_indices.size)
-        n_test_elms = int(self.fraction_test * elm_indices.size)
-
-        test_elms, validation_elms, training_elms = np.split(
-            elm_indices,
-            [n_test_elms, n_test_elms+n_validation_elms]
-        )
-
-        with (self.output_dir/self.data_partition_file).open('w') as data_partition_file:
-            data_partition = {
-                'n_elms': elm_indices.size,
-                'data_location': self.data_location.as_posix(),
-                'training_elms': training_elms.tolist(),
-                'validation_elms': validation_elms.tolist(),
-                'test_elms': test_elms.tolist(),
-            }
-            yaml.safe_dump(
-                data_partition,
-                data_partition_file,
-                default_flow_style=False,
-                sort_keys=False,
-            )
-
-        if not hasattr(self, 'oversample_active_elm'):  # TODO: remove `oversample_active_elm` from base class
-            self.oversample_active_elm = False
-
-        self.raw_label_minmax = []
-
-        self.logger.info(f"Training ELM events: {training_elms.size}")
-        self.train_data = self._preprocess_data(
-            elm_indices=training_elms,
-            shuffle_indices=True,
-            oversample_active_elm=self.oversample_active_elm,
-        )
-
-        if self.fraction_validation > 0.0:
-            self.logger.info(f"Validation ELM events: {validation_elms.size}")
-            self.validation_data = self._preprocess_data(
-                elm_indices=validation_elms,
-                shuffle_indices=False,
-                oversample_active_elm=False,
-                save_filename='validation_elms',
-            )
-        else:
-            self.logger.info("Skipping validation data")
-
-        if self.fraction_test > 0.0:
-            self.logger.info(f"Test ELM events: {test_elms.size}")
-            self.test_data = self._preprocess_data(
-                elm_indices=test_elms,
-                shuffle_indices=False,
-                oversample_active_elm=False,
-                save_filename='test_elms',
-            )
-        else:
-            self.logger.info("Skipping test data")
-
-    def _preprocess_data(
-        self,
-        elm_indices: Iterable = None,
-        shuffle_indices: bool = False,
-        oversample_active_elm: bool = False,
-        save_filename: str = '',
-    ) -> None:
-        packaged_signals = None
-        packaged_window_start = None
-        packaged_valid_t0 = []
-        packaged_labels = []
-        if self.is_regression:
-            label_type = np.float32
-        elif self.is_classification:
-            label_type = np.int8
-        assert label_type
-        if save_filename:
-            plt.ioff()
-            _, axes = plt.subplots(nrows=3, ncols=4, figsize=(16, 9))
-            self.logger.info(f"  Plotting valid indices: {save_filename}.pdf")
-            i_page = 1
-        with h5py.File(self.data_location, 'r') as h5_file:
-            for i_elm, elm_index in enumerate(elm_indices):
-                if save_filename and i_elm%12==0:
-                    for axis in axes.flat:
-                        plt.sca(axis)
-                        plt.cla()
-                elm_key = f"{elm_index:05d}"
-                elm_event = h5_file[elm_key]
-                signals = np.array(elm_event["signals"], dtype=np.float32)  # (64, <time>)
-                signals = np.transpose(signals, (1, 0)).reshape(-1, 8, 8)  # reshape to (<time>, 8, 8)
-                try:
-                    labels = np.array(elm_event["labels"], dtype=label_type)
-                except KeyError:
-                    labels = np.array(elm_event["manual_labels"], dtype=label_type)
-                labels, signals, valid_t0 = self._get_valid_indices(labels, signals)
-                if save_filename:
-                    plt.sca(axes.flat[i_elm%12])
-                    plt.plot(signals[:,2,3]/10, label='BES 20')
-                    plt.plot(signals[:,2,5]/10, label='BES 22')
-                    plt.plot(labels, label='Label')
-                    plt.title(f"ELM index {elm_key}")
-                    plt.legend(fontsize='x-small')
-                    plt.xlabel('Time (mu-s)')
-                    if i_elm%12==11 or i_elm==elm_indices.size-1:
-                        plt.tight_layout()
-                        output_file = self.output_dir/(save_filename + f"_{i_page:02d}.pdf")
-                        plt.savefig(
-                            output_file, 
-                            format="pdf", 
-                            transparent=True,
-                        )
-                        i_page += 1
-                if packaged_signals is None:
-                    packaged_window_start = np.array([0])
-                    packaged_valid_t0 = valid_t0
-                    packaged_signals = signals
-                    packaged_labels = labels
-                else:
-                    last_index = packaged_labels.size - 1
-                    packaged_window_start = np.append(
-                        packaged_window_start, 
-                        last_index + 1
-                    )
-                    packaged_valid_t0 = np.concatenate([packaged_valid_t0, valid_t0])
-                    packaged_signals = np.concatenate([packaged_signals, signals], axis=0)
-                    packaged_labels = np.concatenate([packaged_labels, labels], axis=0)                
-
-        if save_filename:
-            plt.close()
-            pdf_files = sorted(self.output_dir.glob(f'{save_filename}_*.pdf'))
-            output = self.output_dir / f'{save_filename}.pdf'
-            merge_pdfs(pdf_files, output, delete_inputs=True)
-        
-        # valid indices for data sampling
-        packaged_valid_t0_indices = np.arange(packaged_valid_t0.size, dtype="int")
-        packaged_valid_t0_indices = packaged_valid_t0_indices[packaged_valid_t0 == 1]
-
-        if self.is_classification:
-            packaged_valid_t0_indices = self._check_for_balanced_data(
-                packaged_labels=packaged_labels,
-                packaged_valid_t0_indices=packaged_valid_t0_indices,
-                oversample_active_elm=oversample_active_elm,
-            )
-
-        if self.normalize_signals:
-            self.logger.info(f"  Normalizing signals to max ~= 1")
-            packaged_signals /= 10.4  # normalize to max ~= 1
-
-        if self.is_regression and self.normalize_labels:
-            if not self.raw_label_minmax:
-                self.raw_label_minmax = [
-                    packaged_labels.min().item(),
-                    packaged_labels.max().item(),
-                ]
-            self.logger.info(f"  Normalizing labels to min/max = -/+ 1 with raw min/max {self.raw_label_minmax[0]:.4e} {self.raw_label_minmax[1]:.4e}")
-            label_range = self.raw_label_minmax[1] - self.raw_label_minmax[0]
-            packaged_labels = ((packaged_labels - self.raw_label_minmax[0]) / label_range - 0.5) * 2
-
-        if shuffle_indices:
-            self.rng_generator.shuffle(packaged_valid_t0_indices)
-
-        self.logger.info( "  Data tensors -> signals, labels, sample_indices, window_start_indices:")
-        for tensor in [
-            packaged_signals,
-            packaged_labels,
-            packaged_valid_t0_indices,
-            packaged_window_start,
-        ]:
-            tmp = f"  shape {tensor.shape}, dtype {tensor.dtype},"
-            tmp += f" min {tensor.min():.3f}, max {tensor.max():.3f}"
-            if hasattr(tensor, "device"):
-                tmp += f" device {tensor.device[-5:]}"
-            self.logger.info(tmp)
-
-        return_tuple = (
-            packaged_signals, 
-            packaged_labels, 
-            packaged_valid_t0_indices, 
-            packaged_window_start, 
-            elm_indices,
-        )
-        # ensure everything is np.ndarray
-        for item in return_tuple:
-            assert isinstance(item, np.ndarray)
-
-        return return_tuple
-
-    def _check_for_balanced_data(self) -> None:
-        # must implement in subclass
-        raise NotImplementedError
-
-    def _make_datasets(self) -> None:
-        # must implement in subclass
-        raise NotImplementedError
-
-    def _get_valid_indices(self) -> None:
-        # must implement in subclass
-        raise NotImplementedError
-
-    def _make_data_loaders(self) -> None:
-        self.train_data_loader = torch.utils.data.DataLoader(
-                self.train_dataset,
-                batch_size=self.batch_size,
-                shuffle=True if self.seed is None else False,
-                num_workers=self.num_workers,
-                pin_memory=True,
-                drop_last=True,
-            )
-        if self.fraction_validation > 0.0:
-            self.validation_data_loader = torch.utils.data.DataLoader(
-                    self.validation_dataset,
-                    batch_size=self.batch_size,
-                    shuffle=False,
-                    num_workers=self.num_workers,
-                    pin_memory=True,
-                    drop_last=True,
-                )
-
-    def _make_optimizer_scheduler_loss(self) -> None:
-        if self.optimizer_type.lower() == 'adam':
+    def _make_optimizer_scheduler(self) -> None:
+        assert self.optimizer_type in ['adam', 'sgd']
+        if self.optimizer_type == 'adam':
             self.optimizer = torch.optim.Adam(
                 self.model.parameters(), 
                 lr=self.learning_rate, 
                 weight_decay=self.weight_decay,
             )
-        elif self.optimizer_type.lower() == 'sgd':
+        elif self.optimizer_type == 'sgd':
             self.optimizer = torch.optim.SGD(
                 self.model.parameters(), 
                 lr=self.learning_rate, 
@@ -503,13 +198,11 @@ class _Trainer_Base(_Multi_Features_Model_Dataclass):
                 momentum=self.sgd_momentum,
                 dampening=self.sgd_dampening,
             )
-        assert self.optimizer
         self.logger.info(
             f"Optimizer {self.optimizer_type.upper()} " +
             f"with learning rate {self.learning_rate:.1e} " +
             f"and weight decay {self.weight_decay:.1e}"
         )
-
         self.lr_scheduler = torch.optim.lr_scheduler.ReduceLROnPlateau(
             self.optimizer,
             factor=0.5,
@@ -518,26 +211,17 @@ class _Trainer_Base(_Multi_Features_Model_Dataclass):
             verbose=True,
         )
 
+    def _prepare_data(self) -> None:
+        # implement in subclass
+        # must generate self.train_data_loader and self.validation_data_loader
+        raise NotImplementedError
+
     def train(self) -> dict:
         best_score = -np.inf
-        self.results = {
-            'lr': [],
-            'train_loss': [],
-            'valid_loss': [],
-            'scores': [],
-            'scores_label': self.score_function_name,
-            'training_time': [],
-        }
-        checkpoint_file = self.output_dir / self.checkpoint_file
 
+        # TODO: move to ELM classification
         if self.is_classification:
             self.results['roc_scores'] = []
-
-        if self.raw_label_minmax:
-            self.results['raw_label_minmax'] = self.raw_label_minmax
-
-        # send model to device
-        # self.model = self.model.to(self.device)
 
         self.logger.info(f"Batches per epoch {len(self.train_data_loader)}")
         self.logger.info(f"Begin training loop over {self.n_epochs} epochs")
@@ -545,75 +229,54 @@ class _Trainer_Base(_Multi_Features_Model_Dataclass):
         # loop over epochs
         for i_epoch in range(self.n_epochs):
             t_start_epoch = time.time()
-
             self.logger.info(f"Ep {i_epoch+1:03d}: begin")
-            
-            # train_loss = self._train_epoch()
             train_loss = self._single_epoch_loop(
                 is_train=True,
                 data_loader=self.train_data_loader,
             )
-
             self.results['train_loss'].append(train_loss.item())
-
             self.results['lr'].append(self.optimizer.param_groups[0]['lr'])
 
-            score = None
-            valid_loss = None
-            if self.validation_data_loader is not None:
+            if self.validation_data_loader is None:
+                score = None
+                valid_loss = None
+            else:
                 valid_loss, predictions, true_labels = self._single_epoch_loop(
                     is_train=False,
                     data_loader=self.validation_data_loader,
                 )
-
-                self.results['valid_loss'].append(valid_loss.item())
-
-                # apply learning rate scheduler
-                self.lr_scheduler.step(valid_loss)
-
+                self.results['valid_loss'].append(valid_loss.item())  # log validation score
+                self.lr_scheduler.step(valid_loss)  # appy learning rate scheduler
                 if self.is_regression:
-                    score = self.score_function(true_labels, predictions)
+                    score = metrics.r2_score(true_labels, predictions)
                 elif self.is_classification:
                     if self.model.mlp_output_size == 1:
-                        assert hasattr(self, 'threshold')
-                        status = (predictions > self.threshold).astype(int)
-                        score = self.score_function(
-                            true_labels,
-                            status,
-                        )
+                        modified_predictions = (predictions > self.threshold).astype(int)
+                        score = metrics.f1_score(true_labels, modified_predictions, average='binary')
                     else:
-                        status = predictions.argmax(axis=1)
-                        score = self.score_function(
-                            true_labels,
-                            status,
-                            average='weighted'
-                        )
-                score = score.item()
-                self.results['scores'].append(score)
+                        modified_predictions = predictions.argmax(axis=1)
+                        score = metrics.f1_score(true_labels, modified_predictions, average='weighted')
+                self.results['scores'].append(score.item())
 
                 # ROC-AUC score for classification
                 if self.is_classification:
                     if self.model.mlp_output_size == 1:
-                        roc_score = metrics.roc_auc_score(
-                            true_labels,
-                            predictions,
-                        )
+                        roc_score = metrics.roc_auc_score(true_labels, predictions)
                     else:
+                        # TODO: clarify
                         one_hot = np.zeros_like(predictions)
                         for i, j in zip(one_hot, true_labels):
                             i[j] = 1
-                        try:
-                            roc_score = metrics.roc_auc_score(
-                                one_hot,
-                                predictions,
-                                multi_class='ovo', 
-                                average='macro', 
-                                labels=[0, 1, 2, 3],
-                            )
-                        except:
-                            roc_score = np.float32(0)
+                        roc_score = metrics.roc_auc_score(
+                            one_hot,
+                            predictions,
+                            multi_class='ovo', 
+                            average='macro', 
+                            labels=[0, 1, 2, 3],
+                        )
                     self.results['roc_scores'].append(roc_score.item())
 
+            # log training time
             self.results['training_time'].append(time.time()-t_start_training)
 
             # save results to yaml
@@ -629,53 +292,41 @@ class _Trainer_Base(_Multi_Features_Model_Dataclass):
             if score is None or score > best_score:
                 if score is not None:
                     best_score = score
-                    self.logger.info(f"  Best {self.score_function_name}: {best_score:.3f}  saving model...")
-                torch.save(self.model.state_dict(), checkpoint_file.as_posix())
-                self.logger.info(f"  Saved model: {checkpoint_file}  file size: {checkpoint_file.stat().st_size/1e3:.1f} kB")
-                if self.export_onnx:
-                    onnx_file = self.output_dir / 'checkpoint.onnx'
-                    torch.onnx.export(
-                        self.model, 
-                        torch.rand(*self.input_shape)[0].unsqueeze(0),
-                        onnx_file.as_posix(),
-                        input_names=['signal_window'],
-                        output_names=['micro_prediction'],
-                        verbose=True,
-                        opset_version=11
-                    )
-                    self.logger.info(f"  Saved ONNX model: {onnx_file}  file size: {onnx_file.stat().st_size/1e3:.1f} kB")
+                    self.logger.info(f"  Best {self.score_function_name}: {best_score:.3f}")
+                self.model.save_pytorch_model(filename=self.output_dir/self.checkpoint_file)
+                if self.save_onnx_model:
+                    self.model.save_onnx_model(filename=self.output_dir/self.onnx_checkpoint_file)
 
-            # log status
-            status =  f"Ep {i_epoch+1:03d}: "
-            status += f"train loss {train_loss:.3f}  "
+            # print epoch summary
+            modified_predictions =  f"Ep {i_epoch+1:03d}: "
+            modified_predictions += f"train loss {train_loss:.3f}  "
             if score is not None and valid_loss is not None:
-                status += f"val loss {valid_loss:.3f}  "
-                status += f"{self.score_function_name} {score:.3f}  "
+                modified_predictions += f"val loss {valid_loss:.3f}  "
+                modified_predictions += f"{self.score_function_name} {score:.3f}  "
                 if self.is_classification:
-                    status += f"ROC {roc_score:.3f}  "
-            status += f"ep time {time.time()-t_start_epoch:.1f} s "
-            status += f"(total time {time.time()-t_start_training:.1f} s)"
-            self.logger.info(status)
+                    modified_predictions += f"ROC {roc_score:.3f}  "
+            modified_predictions += f"ep time {time.time()-t_start_epoch:.1f} s "
+            modified_predictions += f"(total time {time.time()-t_start_training:.1f} s)"
+            self.logger.info(modified_predictions)
 
-            # report epoch result to optuna
-            if optuna is not None and self.trial is not None:
-                maximize_score = self.trial.user_attrs['maximize_score']
+            # optuna integration; report epoch result to optuna
+            do_prune = False
+            if optuna is not None and self.optuna_trial is not None:
+                maximize_score = self.optuna_trial.user_attrs['maximize_score']
                 if maximize_score is True:
                     assert np.isfinite(score)
-                    self.trial.report(score, i_epoch)
+                    self.optuna_trial.report(score, i_epoch)
                 else:
                     if not np.isfinite(train_loss) or np.isnan(train_loss):
                         train_loss = 1
-                    self.trial.report(train_loss, i_epoch)
-                # save outputs as lists in trial user attributes
+                    self.optuna_trial.report(train_loss, i_epoch)
+                # save results dict in trial user attributes
                 for key in self.results:
-                    self.trial.set_user_attr(key, self.results[key])
-                if self.trial.should_prune():
+                    self.optuna_trial.set_user_attr(key, self.results[key])
+                if self.optuna_trial.should_prune():
+                    do_prune = True
                     self.logger.info("==> Pruning trial with Optuna")
-                    for handler in self.logger.handlers[:]:
-                        handler.close()
-                        self.logger.removeHandler(handler)
-                    optuna.TrialPruned()
+                    break  # exit epoch training loop
 
         self.logger.info(f"End training loop")
         self.logger.info(f"Elapsed time {time.time()-t_start_training:.1f} s")
@@ -683,6 +334,9 @@ class _Trainer_Base(_Multi_Features_Model_Dataclass):
         for handler in self.logger.handlers[:]:
             handler.close()
             self.logger.removeHandler(handler)
+
+        if do_prune:
+            optuna.TrialPruned()
 
         return self.results.copy()
 
@@ -704,7 +358,7 @@ class _Trainer_Base(_Multi_Features_Model_Dataclass):
             mode = 'Valid'
         with context:
             for i_batch, (signal_windows, labels) in enumerate(data_loader):
-                if (i_batch+1) % self.minibatch_interval == 0:
+                if (i_batch+1) % self.minibatch_print_interval == 0:
                     t_start_minibatch = time.time()
                 if is_train:
                     self.optimizer.zero_grad()
@@ -717,8 +371,7 @@ class _Trainer_Base(_Multi_Features_Model_Dataclass):
                     predictions.squeeze(),
                     labels,
                 )
-                if self.is_regression and hasattr(self, 'inverse_weight_loss') and self.inverse_weight_label:
-                    sample_losses = torch.div(sample_losses, labels)
+                sample_losses = self._apply_loss_weight(sample_losses, labels)
                 batch_loss = sample_losses.mean()  # batch loss
                 batch_losses.append(batch_loss.item())  # accumulate batch losses for this epoch
                 if is_train:
@@ -731,7 +384,7 @@ class _Trainer_Base(_Multi_Features_Model_Dataclass):
                     all_predictions.append(predictions.cpu().numpy())
 
                 # minibatch status
-                if (i_batch+1)%self.minibatch_interval == 0:
+                if (i_batch+1)%self.minibatch_print_interval == 0:
                     epoch_loss = np.mean(batch_losses)
                     status =  f"  {mode} batch {i_batch+1:05d}/{len(self.train_data_loader)}  "
                     status += f"batch loss {batch_loss:.3f} (avg loss {epoch_loss:.3f})  "
@@ -750,6 +403,9 @@ class _Trainer_Base(_Multi_Features_Model_Dataclass):
             )
         return return_value
 
+    def _apply_loss_weight(self, losses: torch.Tensor, *args, **kwargs) -> torch.Tensor:
+        return losses
+
 
 if __name__=='__main__':
-    m = _Trainer_Base(dense_num_kernels=8)
+    m = _Base_Trainer(dense_num_kernels=8)
